+++ rpcs3/Emu/CPU/CPUThread.cpp
@@ -1,6 +1,6 @@
 ﻿#include "stdafx.h"
 #include "Emu/System.h"
-#include "Emu/Memory/vm.h"
+#include "Emu/Memory/vm_locking.h"
 #include "CPUThread.h"
 #include "Emu/IdManager.h"
 #include "Utilities/GDBDebugServer.h"
