+++ rpcs3/Emu/Cell/Modules/cellAudioOut.h
@@ -1,6 +1,6 @@
 #pragma once
 
-
+#include "Emu/Memory/vm_ptr.h"
 
 // Error codes
 enum
