+++ rpcs3/Emu/Cell/Modules/cellMsgDialog.h
@@ -1,6 +1,7 @@
 ﻿#pragma once
 
 #include "Utilities/BitField.h"
+#include "Emu/Memory/vm_ptr.h"
 
 enum
 {
