+++ rpcs3/Emu/Cell/PPUThread.h
@@ -1,7 +1,8 @@
 ﻿#pragma once
 
 #include "../CPU/CPUThread.h"
-#include "../Memory/vm.h"
+#include "../Memory/vm_ref.h"
+#include "../Memory/vm_ptr.h"
 #include "Utilities/lockless.h"
 
 enum class ppu_cmd : u32
