+++ rpcs3/Emu/Cell/SPUThread.cpp
@@ -1,7 +1,8 @@
 ﻿#include "stdafx.h"
 #include "Utilities/JIT.h"
 #include "Utilities/sysinfo.h"
-#include "Emu/Memory/vm.h"
+#include "Emu/Memory/vm_ptr.h"
+#include "Emu/Memory/vm_reservation.h"
 #include "Emu/System.h"
 
 #include "Emu/IdManager.h"
