+++ rpcs3/Emu/Cell/lv2/sys_overlay.cpp
@@ -1,5 +1,6 @@
 #include "stdafx.h"
-#include "Emu/Memory/vm.h"
+
+#include "Emu/Memory/vm_ptr.h"
 #include "Emu/System.h"
 #include "Emu/IdManager.h"
 #include "Crypto/unself.h"
