+++ rpcs3/Emu/Cell/lv2/sys_ppu_thread.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "sys_sync.h"
+#include "Emu/Memory/vm_ptr.h"
 
 class ppu_thread;
