+++ rpcs3/Emu/Cell/lv2/sys_process.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "Emu/Memory/vm_ptr.h"
+
 // Process Local Object Type
 enum : u32
 {
