+++ rpcs3/Emu/Cell/lv2/sys_time.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "Emu/Memory/vm_ptr.h"
+
 // SysCalls
 s32 sys_time_get_timezone(vm::ptr<s32> timezone, vm::ptr<s32> summertime);
 s32 sys_time_get_current_time(vm::ptr<s64> sec, vm::ptr<s64> nsec);
