+++ rpcs3/Emu/Cell/lv2/sys_usbd.h
@@ -1,6 +1,6 @@
 #pragma once
 
-
+#include "Emu/Memory/vm_ptr.h"
 
 // SysCalls
