+++ rpcs3/Emu/Memory/vm.h
@@ -3,11 +3,11 @@
 #include <map>
 #include <functional>
 #include <memory>
+#include "Utilities/types.h"
+#include "Utilities/Atomic.h"
 #include "Utilities/VirtualMemory.h"
-
-class shared_mutex;
-class cpu_thread;
-class shared_cond;
+#include "Utilities/StrFmt.h"
+#include "Utilities/BEType.h"
 
 namespace vm
 {
@@ -17,6 +17,8 @@ namespace vm
 	extern u8* const g_stat_addr;
 	extern u8* const g_reservations;
 	extern u8* const g_reservations2;
+
+	struct writer_lock;
 
 	enum memory_location_t : uint
 	{
@@ -47,79 +49,6 @@ namespace vm
 
 	// Address type
 	enum addr_t : u32 {};
-
-	extern shared_mutex g_mutex;
-
-	extern thread_local atomic_t<cpu_thread*>* g_tls_locked;
-
-	// Register reader
-	void passive_lock(cpu_thread& cpu);
-	atomic_t<u64>* passive_lock(const u32 begin, const u32 end);
-
-	// Unregister reader
-	void passive_unlock(cpu_thread& cpu);
-
-	// Unregister reader (foreign thread)
-	void cleanup_unlock(cpu_thread& cpu) noexcept;
-
-	// Optimization (set cpu_flag::memory)
-	void temporary_unlock(cpu_thread& cpu) noexcept;
-	void temporary_unlock() noexcept;
-
-	class reader_lock final
-	{
-		bool m_upgraded = false;
-
-	public:
-		reader_lock(const reader_lock&) = delete;
-		reader_lock& operator=(const reader_lock&) = delete;
-		reader_lock();
-		~reader_lock();
-
-		void upgrade();
-	};
-
-	struct writer_lock final
-	{
-		writer_lock(const writer_lock&) = delete;
-		writer_lock& operator=(const writer_lock&) = delete;
-		writer_lock(u32 addr = 0);
-		~writer_lock();
-	};
-
-	// Get reservation status for further atomic update: last update timestamp
-	inline atomic_t<u64>& reservation_acquire(u32 addr, u32 size)
-	{
-		// Access reservation info: stamp and the lock bit
-		return reinterpret_cast<atomic_t<u64>*>(g_reservations)[addr / 128];
-	}
-
-	// Update reservation status
-	inline void reservation_update(u32 addr, u32 size, bool lsb = false)
-	{
-		// Update reservation info with new timestamp
-		reservation_acquire(addr, size) += 128;
-	}
-
-	// Get reservation sync variable
-	inline shared_cond& reservation_notifier(u32 addr, u32 size)
-	{
-		return *reinterpret_cast<shared_cond*>(g_reservations2 + addr / 128 * 8);
-	}
-
-	void reservation_lock_internal(atomic_t<u64>&);
-
-	inline atomic_t<u64>& reservation_lock(u32 addr, u32 size)
-	{
-		auto& res = vm::reservation_acquire(addr, size);
-
-		if (UNLIKELY(atomic_storage<u64>::bts(res.raw(), 0)))
-		{
-			reservation_lock_internal(res);
-		}
-
-		return res;
-	}
 
 	// Change memory protection of specified memory region
 	bool page_protect(u32 addr, u32 size, u8 flags_test = 0, u8 flags_set = 0, u8 flags_clear = 0);
@@ -172,11 +101,11 @@ namespace vm
 		// Get memory at specified address (if size = 0, addr assumed exact)
 		std::pair<u32, std::shared_ptr<utils::shm>> get(u32 addr, u32 size = 0);
 
-		// Internal
-		u32 imp_used(const vm::writer_lock&);
-
 		// Get allocated memory count
 		u32 used();
+
+		// Internal
+		u32 imp_used(const vm::writer_lock&);
 	};
 
 	// Create new memory block with specified parameters and return it
@@ -343,4 +272,3 @@ namespace vm
 	void close();
 }
 
-#include "vm_var.h"
