+++ rpcs3/Emu/Memory/vm_var.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "vm_ptr.h"
+#include "Utilities/BEType.h"
 
 namespace vm
 {
