+++ rpcs3/Emu/RSX/GCM.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Emu/Memory/vm.h"
+#include "Emu/Memory/vm_ptr.h"
 #include "gcm_enums.h"
 #include "gcm_printing.h"
