+++ rpcs3/Emu/RSX/rsx_methods.cpp
@@ -1,7 +1,7 @@
-﻿#include "stdafx.h"
+#include "stdafx.h"
 #include "rsx_methods.h"
 #include "RSXThread.h"
-#include "Emu/Memory/vm.h"
+#include "Emu/Memory/vm_reservation.h"
 #include "Emu/System.h"
 #include "rsx_utils.h"
 #include "rsx_decode.h"
